Extract cli starter into separate module
