+++ tests/regression_otanalytics.py
@@ -6,8 +6,8 @@ from tqdm import tqdm
 
 from OTAnalytics.application.parser.cli_parser import CliMode
 from OTAnalytics.application.parser.flow_parser import FlowParser
+from OTAnalytics.plugin_cli.cli_application import OtAnalyticsCliApplicationStarter
 from OTAnalytics.plugin_parser.otvision_parser import OtFlowParser
-from OTAnalytics.plugin_ui.main_application import OtAnalyticsCliApplicationStarter
 from tests.utils.assertions import assert_two_files_equal_sorted
 from tests.utils.builders.run_configuration import create_run_config
